Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -1,9 +1,9 @@
 #include <stdio.h>
 int main() {
-    int a,b,c,average;
-    scanf("%d%d%d",&a,&b,&c);
+    double a,b,c,average;
+    scanf("%lf%lf%lf",&a,&b,&c);
     average=a+b+c/3;
-    printf("Average: %.2f\n",average);
+    printf("Average: %.2lf\n",average);
 
     return 0;
 }
